Remove import cycle for model editor queries
